checkstyles corrected
