+++ src/main/java/uk/gov/hmcts/reform/prl/models/dto/cafcass/CafCassCaseData.java
@@ -15,7 +15,7 @@ import uk.gov.hmcts.reform.prl.enums.OrderTypeEnum;
 import uk.gov.hmcts.reform.prl.enums.YesNoDontKnow;
 import uk.gov.hmcts.reform.prl.enums.YesOrNo;
 import uk.gov.hmcts.reform.prl.models.cafcass.hearing.Hearings;
-import uk.gov.hmcts.reform.prl.models.dto.cafcass.manageOrder.OrderCollection;
+import uk.gov.hmcts.reform.prl.models.dto.cafcass.manageorder.CaseOrder;
 
 import java.net.MalformedURLException;
 import java.net.URL;
@@ -215,5 +215,6 @@ public class CafCassCaseData {
     //private List<Element<HearingData>> hearingData;
     private Hearings hearingData;
 
-    public List<OrderCollection> orderCollection;
+    @Setter(AccessLevel.NONE)
+    private List<Element<CaseOrder>> orderCollection;
 }
