--- src/main/java/uk/gov/hmcts/reform/prl/models/dto/cafcass/manageOrder/OrderCollection.java
@@ -1,8 +1,0 @@
-package uk.gov.hmcts.reform.prl.models.dto.cafcass.manageOrder;
-
-public class OrderCollection {
-
-    public String id;
-    public Value value;
-
-}
