--- src/main/java/uk/gov/hmcts/reform/prl/models/dto/cafcass/manageOrder/OrderDocument.java
@@ -1,10 +1,0 @@
-package uk.gov.hmcts.reform.prl.models.dto.cafcass.manageOrder;
-
-public class OrderDocument {
-
-    public Object document_creation_date;
-    public String document_binary_url;
-    public String document_filename;
-    public String document_url;
-
-}
