--- src/main/java/uk/gov/hmcts/reform/prl/models/dto/cafcass/manageOrder/OtherDetails.java
@@ -1,10 +1,0 @@
-package uk.gov.hmcts.reform.prl.models.dto.cafcass.manageOrder;
-
-public class OtherDetails {
-
-    public String createdBy;
-    public String orderRecipients;
-    public String orderCreatedDate;
-    public String orderMadeDate;
-
-}
