--- src/main/java/uk/gov/hmcts/reform/prl/models/dto/cafcass/manageOrder/Value.java
@@ -1,13 +1,0 @@
-package uk.gov.hmcts.reform.prl.models.dto.cafcass.manageOrder;
-
-import java.util.Date;
-
-public class Value {
-
-    public String orderType;
-    public OrderDocument orderDocument;
-    public Date dateCreated;
-    public OtherDetails otherDetails;
-    public String orderTypeId;
-
-}
