+++ src/main/java/uk/gov/hmcts/reform/prl/models/dto/cafcass/manageorder/OrderCollection.java
@@ -1,0 +1,21 @@
+package uk.gov.hmcts.reform.prl.models.dto.cafcass.manageorder;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@Builder(toBuilder = true)
+public class OrderCollection {
+
+    public String id;
+    public CaseOrder caseOrder;
+
+}
