+++ src/main/java/uk/gov/hmcts/reform/prl/models/dto/cafcass/manageorder/OrderDocument.java
@@ -1,0 +1,26 @@
+package uk.gov.hmcts.reform.prl.models.dto.cafcass.manageorder;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@Builder(toBuilder = true)
+public class OrderDocument {
+
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
+    public String documentCreationDate;
+
+    public String documentBinaryUrl;
+    public String documentFilename;
+    public String documentUrl;
+
+}
